kademlia fixes (#266)

+++ src/protocol/kademlia/impl/peer_routing_table_impl.cpp
@@ -27,15 +27,6 @@ OUTCOME_CPP_DEFINE_CATEGORY(libp2p::protocol::kademlia,
   return "unknown error";
 }
 
-namespace {
-  using libp2p::protocol::kademlia::Bucket;
-
-  size_t getBucketId(const std::vector<Bucket> &buckets, size_t cpl) {
-    BOOST_ASSERT(not buckets.empty());
-    return cpl >= buckets.size() ? buckets.size() - 1 : cpl;
-  }
-}  // namespace
-
 namespace libp2p::protocol::kademlia {
 
   size_t Bucket::size() const {
@@ -47,19 +38,23 @@ namespace libp2p::protocol::kademlia {
   }
 
   void Bucket::sort(const NodeId &node_id) {
-    XorDistanceComparator cmp(node_id);
+    XorDistanceComparator cmp{node_id};
     peers_.sort(cmp);
   }
 
+  auto findPeer(auto &peers, const peer::PeerId &p) {
+    return std::ranges::find_if(peers,
+                                [&p](const auto &i) { return i.peer_id == p; });
+  }
+
   auto Bucket::find(const peer::PeerId &p) const {
-    return std::find_if(peers_.begin(), peers_.end(), [&p](const auto &i) {
-      return i.peer_id == p;
-    });
+    return findPeer(peers_, p);
   }
 
   bool Bucket::moveToFront(const PeerId &pid) {
-    auto it = find(pid);
+    auto it = findPeer(peers_, pid);
     if (it != peers_.end()) {
+      it->is_connected = true;
       if (it != peers_.begin()) {
         peers_.splice(peers_.begin(), peers_, it);
       }
@@ -68,15 +63,18 @@ namespace libp2p::protocol::kademlia {
     return true;
   }
 
-  void Bucket::emplaceToFront(const PeerId &pid, bool is_replaceable) {
-    peers_.emplace(peers_.begin(), pid, is_replaceable);
+  void Bucket::emplaceToFront(const PeerId &pid,
+                              bool is_replaceable,
+                              bool is_connected) {
+    peers_.emplace(peers_.begin(), pid, is_replaceable, is_connected);
   }
 
   boost::optional<PeerId> Bucket::removeReplaceableItem() {
     boost::optional<PeerId> result;
 
     for (auto it = peers_.rbegin(); it != peers_.rend(); ++it) {
-      if (it->is_replaceable) {
+      // https://github.com/libp2p/rust-libp2p/blob/3837e33cd4c40ae703138e6aed6f6c9d52928a80/protocols/kad/src/kbucket/bucket.rs#L310-L366
+      if (it->is_replaceable and not it->is_connected) {
         result = std::move(it->peer_id);
         peers_.erase((++it).base());
         break;
@@ -119,25 +117,6 @@ namespace libp2p::protocol::kademlia {
     return false;
   }
 
-  Bucket Bucket::split(size_t commonLenPrefix, const NodeId &target) {
-    Bucket b{};
-
-    std::list<BucketPeerInfo> new_peers;
-
-    while (!peers_.empty()) {
-      auto it = peers_.begin();
-      if (it->node_id.commonPrefixLen(target) > commonLenPrefix) {
-        b.peers_.splice(b.peers_.end(), peers_, it);
-      } else {
-        new_peers.splice(new_peers.end(), peers_, it);
-      }
-    }
-
-    peers_.swap(new_peers);
-
-    return b;
-  }
-
   PeerRoutingTableImpl::PeerRoutingTableImpl(
       const Config &config,
       std::shared_ptr<peer::IdentityManager> identity_manager,
@@ -148,55 +127,65 @@ namespace libp2p::protocol::kademlia {
         local_([this] {
           BOOST_ASSERT(identity_manager_ != nullptr);
           return identity_manager_->getId();
-        }()),
-        log_("PeerRoutingTable", "kademlia") {
+        }()) {
     BOOST_ASSERT(identity_manager_ != nullptr);
     BOOST_ASSERT(bus_ != nullptr);
     BOOST_ASSERT(config_.maxBucketSize > 1);
-    buckets_.emplace_back();  // create initial bucket
   }
 
   void PeerRoutingTableImpl::remove(const peer::PeerId &peer_id) {
-    size_t cpl = NodeId(peer_id).commonPrefixLen(local_);
-    auto bucket_id = getBucketId(buckets_, cpl);
-    auto &bucket = buckets_.at(bucket_id);
+    auto bucket_index = getBucketIndex(NodeId{peer_id});
+    if (not bucket_index) {
+      return;
+    }
+    auto &bucket = buckets_.at(*bucket_index);
     bucket.remove(peer_id);
     bus_->getChannel<event::protocol::kademlia::PeerRemovedChannel>().publish(
         peer_id);
   }
 
+  // https://github.com/libp2p/rust-libp2p/blob/3837e33cd4c40ae703138e6aed6f6c9d52928a80/protocols/kad/src/kbucket.rs#L400-L461
   std::vector<peer::PeerId> PeerRoutingTableImpl::getNearestPeers(
       const NodeId &node_id, size_t count) {
-    size_t cpl = node_id.commonPrefixLen(local_);
-    size_t bucketId = getBucketId(buckets_, cpl);
-
-    // make a copy of a bucket
-    auto bucket = buckets_.at(bucketId);
-    if (bucket.size() < count) {
-      // In the case of an unusual split, one bucket may be short or empty.
-      // if this happens, search both surrounding buckets for nearby peers
-      if (bucketId > 0) {
-        auto &left = buckets_.at(bucketId - 1);
-        bucket.append(left);
-      }
-      if (bucketId < buckets_.size() - 1) {
-        auto &right = buckets_.at(bucketId + 1);
-        bucket.append(right);
+    auto distance = local_.distance(node_id);
+    // check if distance bit is set
+    // https://github.com/libp2p/rust-libp2p/blob/3837e33cd4c40ae703138e6aed6f6c9d52928a80/protocols/kad/src/kbucket.rs#L409-L428
+    auto bit = [&](size_t i) {
+      auto j = 255 - i;
+      return ((distance[j / 8] >> (7 - j % 8)) & 1) != 0;
+    };
+    auto bucket_index = getBucketIndex(node_id);
+    Bucket result;
+    auto done = [&] { return result.size() >= count; };
+    auto append = [&](size_t i) { result.append(buckets_.at(i)); };
+    if (bucket_index) {
+      if (auto i = *bucket_index) {
+        append(i);
+        for (--i; i != 0 and not done(); --i) {
+          if (bit(i)) {
+            append(i);
+          }
+        }
       }
     }
-
-    // sort bucket in ascending order by XOR distance from local peer.
-    bucket.sort(node_id);
-
-    bucket.truncate(count);
-
-    return bucket.peerIds();
+    if (not done()) {
+      append(0);
+    }
+    for (size_t i = 1; i < 256 and not done(); ++i) {
+      if (not bit(i)) {
+        append(i);
+      }
+    }
+    result.sort(node_id);
+    result.truncate(count);
+    return result.peerIds();
   }
 
   namespace {
     outcome::result<bool> replacePeer(Bucket &bucket,
                                       const peer::PeerId &pid,
                                       bool is_replaceable,
+                                      bool is_connected,
                                       event::Bus &bus) {
       const auto removed = bucket.removeReplaceableItem();
       if (!removed.has_value()) {
@@ -204,7 +193,7 @@ namespace libp2p::protocol::kademlia {
       }
       bus.getChannel<event::protocol::kademlia::PeerRemovedChannel>().publish(
           removed.value());
-      bucket.emplaceToFront(pid, is_replaceable);
+      bucket.emplaceToFront(pid, is_replaceable, is_connected);
       bus.getChannel<event::protocol::kademlia::PeerAddedChannel>().publish(
           pid);
       return true;
@@ -214,11 +203,11 @@ namespace libp2p::protocol::kademlia {
   outcome::result<bool> PeerRoutingTableImpl::update(const peer::PeerId &pid,
                                                      bool is_permanent,
                                                      bool is_connected) {
-    NodeId nodeId(pid);
-    size_t cpl = nodeId.commonPrefixLen(local_);
-
-    auto bucketId = getBucketId(buckets_, cpl);
-    auto &bucket = buckets_.at(bucketId);
+    auto bucket_index = getBucketIndex(NodeId{pid});
+    if (not bucket_index) {
+      return true;
+    }
+    auto &bucket = buckets_.at(*bucket_index);
 
     // Trying to find and move to front if its a long lived connected peer
     if (is_connected) {
@@ -230,52 +219,13 @@ namespace libp2p::protocol::kademlia {
     }
 
     if (bucket.size() < config_.maxBucketSize) {
-      bucket.emplaceToFront(pid, !is_permanent);
+      bucket.emplaceToFront(pid, not is_permanent, is_connected);
       bus_->getChannel<event::protocol::kademlia::PeerAddedChannel>().publish(
           pid);
       return true;
     }
 
-    if (bucketId == buckets_.size() - 1) {
-      // this is last bucket
-      // if the bucket is too large and this is the last bucket (i.e. wildcard),
-      // unfold it.
-      nextBucket();
-
-      // the structure of the table has changed, so let's recheck if the peer
-      // now has a dedicated bucket.
-      auto resizedBucketId = getBucketId(buckets_, cpl);
-      auto &resizedBucket = buckets_.at(resizedBucketId);
-      if (resizedBucket.size() < config_.maxBucketSize) {
-        resizedBucket.emplaceToFront(pid, !is_permanent);
-        bus_->getChannel<event::protocol::kademlia::PeerAddedChannel>().publish(
-            pid);
-        return true;
-      }
-
-      return replacePeer(resizedBucket, pid, !is_permanent, *bus_);
-    }
-
-    return replacePeer(bucket, pid, !is_permanent, *bus_);
-  }
-
-  void PeerRoutingTableImpl::nextBucket() {
-    // This is the last bucket, which allegedly is a mixed bag containing
-    // peers not belonging in dedicated (unfolded) buckets. _allegedly_ is
-    // used here to denote that *all* peers in the last bucket might feasibly
-    // belong to another bucket. This could happen if e.g. we've unfolded 4
-    // buckets, and all peers in folded bucket 5 really belong in bucket 8.
-
-    // get last bucket
-    auto bucket = buckets_.rbegin();
-    // split it
-    auto newBucket = bucket->split(buckets_.size() - 1, local_);
-    buckets_.push_back(std::move(newBucket));
-
-    auto newLastBucket = buckets_.rbegin();
-    if (newLastBucket->size() > config_.maxBucketSize) {
-      nextBucket();
-    }
+    return replacePeer(bucket, pid, not is_permanent, is_connected, *bus_);
   }
 
   size_t PeerRoutingTableImpl::size() const {
@@ -293,5 +243,14 @@ namespace libp2p::protocol::kademlia {
       vec.insert(vec.end(), peer_ids.begin(), peer_ids.end());
     }
     return vec;
+  }
+
+  // https://github.com/libp2p/rust-libp2p/blob/3837e33cd4c40ae703138e6aed6f6c9d52928a80/protocols/kad/src/kbucket.rs#L141-L150
+  std::optional<size_t> PeerRoutingTableImpl::getBucketIndex(
+      const NodeId &key) const {
+    if (local_ == key) {
+      return std::nullopt;
+    }
+    return 255 - local_.commonPrefixLen(key);
   }
 }  // namespace libp2p::protocol::kademlia
